+++ tests/BaseTest.php
@@ -1,8 +1,14 @@
 <?php namespace STPH\pdfInjector;
 
+use \ExternalModules\ExternalModules;
+use \Exception;
+
 require_once __DIR__ . '/../../../redcap_connect.php';
 
 abstract class BaseTest extends \ExternalModules\ModuleBaseTest {
+
+    static string $pid1;
+    static string $pid2;
 
     protected function callPrivateMethod($name, array $args) {
         $obj =  $this->module;
@@ -10,6 +16,104 @@ abstract class BaseTest extends \ExternalModules\ModuleBaseTest {
         $method = $class->getMethod($name);
         $method->setAccessible(true);
         return $method->invokeArgs($obj, $args);
+    }
+
+    static function setUpBeforeClass():void{
+
+        if(!self::hasTestProjects()) {
+
+            self::setupTestProjects();
+
+        } else {
+
+            $testPIDs = ExternalModules::getTestPIDs();
+            self::$pid1 = $testPIDs[0];
+            self::$pid2 = $testPIDs[1];    
+        }
+
+
+        //  Fake Project ID (using first test project)
+        define(PROJECT_ID, ExternalModules::getTestPIDs()[0]);
+
+    }
+
+    
+    //  Check if system has test projects
+    private static function hasTestProjects(){
+
+        $sql = "SELECT *  FROM redcap_config WHERE field_name = 'external_modules_test_pids'";
+        $result  = ExternalModules::query($sql, []);
+
+        if($result->num_rows > 0) {
+            return true;
+        } else {
+            return false;
+        }
+
+    }
+
+    //  Setup test projects
+    static function setupTestProjects() {
+
+        self::$pid1 = self::createTestProject('External Module Unit Test Project 1');
+        self::$pid2 = self::createTestProject('External Module Unit Test Project 2');
+        
+        $value = self::$pid1.",".self::$pid2;
+
+        $sql = "insert into redcap_config values ('external_modules_test_pids', ?)";
+        
+        $result = ExternalModules::query($sql, [$value]);
+    }
+
+    
+    //  Create test project
+    static function createTestProject($title){
+
+        $title = \Project::cleanTitle($title);
+        $new_app_name = \Project::getValidProjectName($title);
+
+        $purpose = 0;
+        $project_note = 'Unit Testing';
+
+        $ui_id = NULL;
+        $auto_inc_set = 1;
+        $GLOBALS['__SALT__'] = substr(sha1(rand()), 0, 10);
+
+        ExternalModules::query("insert into redcap_projects (project_name, purpose, app_title, creation_time, created_by, auto_inc_set, project_note,auth_meth,__SALT__) values(?,?,?,?,?,?,?,?,?)",
+            [$new_app_name,$purpose,$title,NOW,$ui_id,$auto_inc_set,trim($project_note),'none',$GLOBALS['__SALT__']]);            
+
+        // Get this new project's project_id
+        $pid = db_insert_id();
+
+        // Insert project defaults into redcap_projects
+        \Project::setDefaults($pid);
+
+        return $pid;
+
+    }
+        
+    static function tearDownAfterClass():void{
+        //self::cleanupTestProjects();    
+    }
+
+
+    //  Cleanup test projects
+    static function cleanupTestProjects() {
+        ExternalModules::query(
+            "DELETE FROM `redcap_projects` WHERE `project_id`= ? OR `project_id`=?", 
+            [
+                self::$pid1,
+                self::$pid2
+            ]
+        );
+
+        ExternalModules::query(
+            "DELETE FROM `redcap_config` WHERE  `field_name`='external_modules_test_pids'", []
+        );
+    }    
+
+    function getFirstTestPID(){
+        return ExternalModules::getTestPIDs()[0];
     }
 
 }
